requirement.txt
streamlit
panda
numpy
scikit-learn
pickle
python-dotenv
ast